+++ py/desitarget/sv1/sv1_cuts.py
@@ -25,7 +25,8 @@ from astropy.coordinates import SkyCoord
 
 from desitarget.cuts import _getColors, _psflike, _check_BGS_targtype_sv
 from desitarget.cuts import shift_photo_north
-from desitarget.gaiamatch import is_in_Galaxy, find_gaia_files_hp, gaia_psflike
+from desitarget.gaiamatch import is_in_Galaxy, find_gaia_files_hp
+from desitarget.gaiamatch import gaia_psflike, unextinct_gaia_mags
 from desitarget.geomask import imaging_mask
 
 # ADM set up the DESI default logger
@@ -37,7 +38,7 @@ start = time()
 
 
 def isGAIA_STD(ra=None, dec=None, galb=None, gaiaaen=None, pmra=None, pmdec=None,
-               parallax=None, parallaxovererror=None, gaiabprpfactor=None,
+               parallax=None, parallaxovererror=None, ebv=None, gaiabprpfactor=None,
                gaiasigma5dmax=None, gaiagmag=None, gaiabmag=None, gaiarmag=None,
                gaiadupsource=None, gaiaparamssolved=None,
                primary=None, test=False, nside=2):
@@ -45,6 +46,8 @@ def isGAIA_STD(ra=None, dec=None, galb=None, gaiaaen=None, pmra=None, pmdec=None
 
     Parameters
     ----------
+    ebv : :class:`array_like`
+        E(B-V) values from the SFD dust maps.
     test : :class:`bool`, optional, defaults to ``False``
         If ``True``, then we're running unit tests and don't have to
         find and read every possible Gaia file.
@@ -64,7 +67,7 @@ def isGAIA_STD(ra=None, dec=None, galb=None, gaiaaen=None, pmra=None, pmdec=None
     Notes
     -----
     - See :func:`~desitarget.cuts.set_target_bits` for other parameters.
-    - Current version (01/15/21) is version 151 on `the SV wiki`_.
+    - Current version (01/28/21) is version 154 on `the SV wiki`_.
     """
     if primary is None:
         primary = np.ones_like(gaiagmag, dtype='?')
@@ -96,9 +99,12 @@ def isGAIA_STD(ra=None, dec=None, galb=None, gaiaaen=None, pmra=None, pmdec=None
     ispsf = gaia_psflike(gaiaaen, gaiagmag)
     std &= ispsf
 
+    # ADM de-extinct the magnitudes before applying color cuts.
+    gd, bd, rd = unextinct_gaia_mags(gaiagmag, gaiabmag, gaiarmag, ebv)
+
     # ADM apply the Gaia color cuts for standards.
-    bprp = gaiabmag - gaiarmag
-    gbp = gaiagmag - gaiabmag
+    bprp = bd - rd
+    gbp = gd - bd
     std &= bprp > 0.2
     std &= bprp < 0.9
     std &= gbp > -1.*bprp/2.0