+++ doc/changes.rst
@@ -5,12 +5,25 @@ desitarget Change Log
 0.49.1 (unreleased)
 -------------------
 
+* Improved functionality for Gaia-only targets [`PR #673`_]. Includes:
+    * Deredden Gaia colors when selecting Gaia-only standards.
+        * Add a general function to de-extinct Gaia magnitudes.
+        * Add E(B-V) to supplemental targets for downstream calibrations.
+    * Add a ``flavor`` property to the secondary targets.
+        * Flavor can be "SPARE", "DEDICATED", "SSV", "QSO", "TOO".
+        * Add unit test to check allowed values of flavor.
+    * Update priorities for dedicated secondary targets.
+        * Now they can be distinguished from spare targets by ``flavor``.
+    * Include ``PHOTSYS`` of "G" for Gaia-only targets.
+    * Fix minor bugs that arose due to recent changes in the data model.
+    * Add code to scrape Gaia EDR3 flat (csv) files from the web.
+        * And reformat them as FITS files, and by HEALPixel.
 * Updates to MWS main survey target cuts [`PR #672`_]. Includes:
-  * Add bright limit to MWS_NEARBY
-  * Add MWS_BHB as main survey class
+    * Add bright limit to MWS_NEARBY
+    * Add MWS_BHB as main survey class
 
 .. _`PR #672`: https://github.com/desihub/desitarget/pull/672
-
+.. _`PR #673`: https://github.com/desihub/desitarget/pull/673
 
 0.49.0 (2020-01-18)
 -------------------